modify to compile on msvc

+++ src/memory.h
@@ -1,9 +1,9 @@
-extern void *chl_alloc (long nbytes);
+extern void *chl_alloc (size_t nbytes);
 
-extern void *chl_calloc (long count, long nbytes);
+extern void *chl_calloc (size_t count, size_t nbytes);
 
 extern void chl_free (void *ptr);
 
 #define ALLOC(nbytes) chl_alloc ((nbytes))
-#define NEW(p) ((p) = ALLOC ((long) sizeof *(p)))
+#define NEW(p) ((p) = ALLOC (sizeof *(p)))
 #define FREE(ptr) ((void) (chl_free ((ptr)), (ptr) = 0))
